+++ Lib/menus/variable/openVarFontAssistant_old.py
@@ -1,0 +1,10 @@
+# menuTitle: VarFontAssistant (old)
+
+from importlib import reload
+import xTools4.dialogs.variable.old.VarFontAssistant
+reload(xTools4.dialogs.variable.old.VarFontAssistant)
+
+from mojo.roboFont import OpenWindow
+from xTools4.dialogs.variable.old.VarFontAssistant import VarFontAssistant
+
+OpenWindow(VarFontAssistant)