+++ docs/changelog.md
@@ -34,6 +34,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), 
 
 - - -
 {: .mb-4 }
+ 
+
+<!--
+
+- adding validation colors to attributes in VarGlyphAssistant, removing points tab
+- 
+
+-->
 
 
 0.1.4